refactor: Fix circular dependency with storage callbacks
