+++ packages/core/echo/indexing/package.json
@@ -18,12 +18,8 @@
   "dependencies": {
     "@dxos/async": "workspace:*",
     "@dxos/automerge": "workspace:*",
-    "@dxos/codec-protobuf": "workspace:*",
     "@dxos/context": "workspace:*",
-    "@dxos/debug": "workspace:*",
-    "@dxos/echo-pipeline": "workspace:*",
     "@dxos/echo-protocol": "workspace:*",
-    "@dxos/echo-schema": "workspace:*",
     "@dxos/invariant": "workspace:*",
     "@dxos/keys": "workspace:*",
     "@dxos/kv-store": "workspace:*",
@@ -34,8 +30,6 @@
     "@dxos/tracing": "workspace:*",
     "@dxos/util": "workspace:*",
     "@orama/orama": "^2.0.8",
-    "abstract-level": "^1.0.2",
-    "level": "^8.0.1",
     "level-transcoder": "^1.0.1",
     "lodash.isequal": "^4.5.0"
   },
