+++ packages/core/echo/indexing/src/index.ts
@@ -6,5 +6,4 @@ export * from './indexer';
 export * from './index-metadata-store';
 export * from './index-schema';
 export * from './index-store';
-export * from './storage-callbacks';
 export * from './types';
